refactor(split-rendering): extract leaf modules (Phase 1)
Extract four self-contained leaf modules from split_rendering:

- spans.rs: push_span_with_map, SpanAccumulator, DebugSpanTracker,
  span_color_at/info_at, compress_chars, compute_inline_diff,
  debug_tag_style, push_debug_tag
- style.rs: inline_diagnostic_style, fold_placeholder_style,
  dim_color_for_tilde, append_fold_placeholder, create_virtual_line
- char_style.rs: compute_char_style with private CharStyleContext/Output
- post_pass.rs: render_column_guides, render_ruler_bg,
  apply_hyperlink_overlays, apply_osc8_to_cells, apply_background_to_lines

None of these modules depend on any shared render-time "mega struct".
All 37 split_rendering tests pass.
